gcx: fix mmu context switch issues.
On an mmu context switch, an mmu flush is required.

When the mmu is enabled after a power off, the mmu context
at the head of the cmd buffer queue is made current but the
cached current context was not updated so the next process
to try a commit would throw an l3 bus error.

Change-Id: I21414576c574d5cdeebd917f5b02d8d1e4ce3115
Signed-off-by: Craig Stout <craig.stout@ti.com>